+++ plugins/aea-ledger-ethereum/aea_ledger_ethereum/ethereum.py
@@ -63,6 +63,9 @@ DEFAULT_CURRENCY_DENOM = "wei"
 ETH_GASSTATION_URL = "https://ethgasstation.info/api/ethgasAPI.json"
 _ABI = "abi"
 _BYTECODE = "bytecode"
+EIP1559 = "eip1559"
+GAS_STATION = "gas_station"
+AVAILABLE_STRATEGIES = (EIP1559, GAS_STATION)
 
 MAX_GAS_FAST = 1500
 
@@ -100,6 +103,14 @@ DEFAULT_EIP1559_STRATEGY = {
 
 
 DEFAULT_GAS_STATION_STRATEGY = {"gas_price_api_key": "", "gas_price_strategy": "fast"}
+
+DEFAULT_GAS_PRICE_STRATEGIES = {
+    EIP1559: DEFAULT_EIP1559_STRATEGY,
+    GAS_STATION: DEFAULT_GAS_STATION_STRATEGY,
+}
+
+# The tip increase is the minimum required of 10%.
+TIP_INCREASE = 1.1
 
 
 def wei_to_gwei(number: Type[int]) -> Union[int, decimal.Decimal]:
@@ -185,9 +196,11 @@ def get_gas_price_strategy_eip1559(
         transaction_params: TxParams,  # pylint: disable=unused-argument
     ) -> Dict[str, Wei]:
         """
-        Get gas price from Eth Gas Station api.
+        Get gas price using EIP1559.
 
-        Visit `https://docs.ethgasstation.info/gas-price` for documentation.
+        Visit `https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1559.md`
+        for more information.
+
         :param web3: web3 instance
         :param transaction_params: transaction parameters
         :return: dictionary containing gas price strategy
@@ -742,8 +755,8 @@ class EthereumApi(LedgerApi, EthereumHelper):
     identifier = _ETHEREUM
 
     _gas_price_strategy_callables: Dict[str, Callable] = {
-        "gas_station": get_gas_price_strategy,
-        "eip1559": get_gas_price_strategy_eip1559,
+        GAS_STATION: get_gas_price_strategy,
+        EIP1559: get_gas_price_strategy_eip1559,
     }
 
     def __init__(self, **kwargs: Any):
@@ -759,10 +772,15 @@ class EthereumApi(LedgerApi, EthereumHelper):
         self._is_gas_estimation_enabled = kwargs.pop("is_gas_estimation_enabled", False)
 
         self._default_gas_price_strategy: str = kwargs.pop(
-            "default_gas_price_strategy", "eip1559"
+            "default_gas_price_strategy", EIP1559
         )
+        if self._default_gas_price_strategy not in AVAILABLE_STRATEGIES:
+            raise ValueError(
+                f"Gas price strategy must be one of {AVAILABLE_STRATEGIES}, provided: {self._default_gas_price_strategy}"
+            )  # pragma: nocover
+
         self._gas_price_strategies: Dict[str, Dict] = kwargs.pop(
-            "gas_price_strategies", {}
+            "gas_price_strategies", DEFAULT_GAS_PRICE_STRATEGIES
         )
 
     @property
@@ -874,7 +892,7 @@ class EthereumApi(LedgerApi, EthereumHelper):
             transaction.update({"gasPrice": gas_price})
 
         if gas_price is None and max_fee_per_gas is None:
-            gas_pricing = self._try_get_gas_pricing(
+            gas_pricing = self.try_get_gas_pricing(
                 gas_price_strategy, gas_price_strategy_extra_config
             )
             if gas_pricing is None:
@@ -883,40 +901,89 @@ class EthereumApi(LedgerApi, EthereumHelper):
 
         return transaction
 
-    @try_decorator("Unable to retrieve gas price: {}", logger_method="warning")
-    def _try_get_gas_pricing(
+    def _get_gas_price_strategy_callable(
         self,
         gas_price_strategy: Optional[str] = None,
         extra_config: Optional[Dict] = None,
-    ) -> Optional[Dict[str, int]]:
-        """Try get the gas price based on the provided strategy."""
+    ) -> Callable:
+        """
+        Returns parameters for gas price callable.
 
+        Note: The priority of gas price callable will be
+        `extra_config(Runtime params) > self._gas_price_strategies (Set using config file.) > DEFAULT_GAS_PRICE_STRATEGIES (Default values.)`
+
+        :param gas_price_strategy: name of the gas price strategy.
+        :param extra_config: gas price strategy getter parameters.
+        :return: gas price callable.
+        """
         gas_price_strategy = (
             gas_price_strategy
             if gas_price_strategy is not None
             else self._default_gas_price_strategy
         )
+        if gas_price_strategy not in AVAILABLE_STRATEGIES:  # pragma: nocover
+            _default_logger.debug(
+                f"Gas price strategy must be one of {AVAILABLE_STRATEGIES}, provided: {self._default_gas_price_strategy}"
+            )
+            return None
+
         _default_logger.debug(f"Using strategy: {gas_price_strategy}")
-        extra_config = extra_config or {}
         gas_price_strategy_getter = self._gas_price_strategy_callables.get(
             gas_price_strategy, None
         )
-        if gas_price_strategy_getter is None:
-            _default_logger.debug(
-                "No strategy found! Check configuration of ethereum-ledger-api."
-            )
-            return None  # pragma: nocover
-        kwargs = self._gas_price_strategies.get(gas_price_strategy, {}).copy()
-        kwargs.update(extra_config)
-        gas_price_strategy_callable = gas_price_strategy_getter(**kwargs)
+
+        parameters = DEFAULT_GAS_PRICE_STRATEGIES.get(gas_price_strategy)
+        parameters.update(self._gas_price_strategies.get(gas_price_strategy, {}))
+        parameters.update(extra_config or {})
+        return gas_price_strategy_getter(**parameters)
+
+    @try_decorator("Unable to retrieve gas price: {}", logger_method="warning")
+    def try_get_gas_pricing(
+        self,
+        gas_price_strategy: Optional[str] = None,
+        extra_config: Optional[Dict] = None,
+        old_tip: Optional[int] = None,
+    ) -> Optional[Dict[str, int]]:
+        """
+        Try get the gas price based on the provided strategy.
+
+        :param gas_price_strategy: the gas price strategy to use, e.g., the EIP-1559 strategy.
+            Can be either `eip1559` or `gas_station`.
+        :param extra_config: gas price strategy getter parameters.
+        :param old_tip: the old `maxPriorityFeePerGas` in case that we are trying to resubmit a transaction.
+        :return: a dictionary with the gas data.
+        """
+
+        gas_price_strategy_callable = self._get_gas_price_strategy_callable(
+            gas_price_strategy, extra_config,
+        )
+        if gas_price_strategy_callable is None:  # pragma: nocover
+            return None
 
         prior_strategy = self._api.eth.gasPriceStrategy
         try:
             self._api.eth.set_gas_price_strategy(gas_price_strategy_callable)
             gas_price = self._api.eth.generate_gas_price()
         finally:
-            if prior_strategy is not None:
-                self._api.eth.set_gas_price_strategy(prior_strategy)  # pragma: nocover
+            self._api.eth.set_gas_price_strategy(prior_strategy)  # pragma: nocover
+
+        if old_tip is not None and gas_price_strategy is EIP1559:
+            base_fee_per_gas = (
+                gas_price["maxFeePerGas"] - gas_price["maxPriorityFeePerGas"]
+            )
+            updated_max_priority_fee_per_gas = old_tip * TIP_INCREASE
+            updated_max_fee_per_gas = (
+                updated_max_priority_fee_per_gas + base_fee_per_gas
+            )
+
+            if gas_price["maxPriorityFeePerGas"] < updated_max_priority_fee_per_gas:
+                gas_price["maxPriorityFeePerGas"] = updated_max_priority_fee_per_gas
+                gas_price["maxFeePerGas"] = updated_max_fee_per_gas
+
+        elif old_tip is not None and gas_price_strategy is GAS_STATION:
+            updated_gas_price = old_tip * TIP_INCREASE
+            gas_price["gasPrice"] = max(gas_price["gasPrice"], updated_gas_price)
+
         return gas_price
 
     @try_decorator("Unable to retrieve transaction count: {}", logger_method="warning")
@@ -1103,15 +1170,17 @@ class EthereumApi(LedgerApi, EthereumHelper):
             transaction.update({"gasPrice": gas_price})
 
         if gas_price is None and max_fee_per_gas is None:
-            gas_pricing = self._try_get_gas_pricing(
+            gas_pricing = self.try_get_gas_pricing(
                 gas_price_strategy, gas_price_strategy_extra_config
             )
 
             if gas_pricing is None:
                 return None  # pragma: nocover
+
             transaction.update(gas_pricing)
 
         transaction = instance.constructor(**kwargs).buildTransaction(transaction)
+
         if transaction is None:
             return None  # pragma: nocover
         transaction.pop("to", None)  # only 'from' address, don't insert 'to' address!